Various issues fixed and ngx audio player replaced
Issue with You May Like component not changing release details when artwork clicked has been resolved. Activated route subscription added to component init to ensure new data is passed to the component whenver the id changes on the URL.

Bandcamp embedded player replaced ngx-audio player as was unable to dynamically update the array that the player required for the track title and link. Release details component now working as I want it to.